pyproject.toml
[tool.poetry]
name = "spotify-web"
version = "0.1.0"
description = "A web app using spotify"
authors = ["No767, others"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
Flask = "^2.2.2"
blinker = "^1.5"
python-dotenv = "^0.21.0"
watchdog = "^2.1.9"
Django = "^4.1.1"
Flask-Session = "^0.4.0"
spotipy = "^2.20.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
